PKX-244 User Registration API
